+++ packages/expo-router/src/native-tabs/__rsc_tests__/tabs.test.tsx
@@ -4,7 +4,6 @@ import * as React from 'react';
 // Right now NativeTabs.Trigger.* components cannot be used in the server environment
 import { Badge, Icon, Label } from '../../primitives';
 import { NativeTabs } from '../NativeTabs';
-import { NativeTabsTriggerTabBar } from '../NativeTabsTriggerTabBar';
 
 it(`renders NativeTabs`, async () => {
   await expect(<NativeTabs />).toMatchFlightSnapshot();
@@ -32,8 +31,4 @@ it(`renders Badge`, async () => {
 
 it(`renders Label`, async () => {
   await expect(<Label>Test</Label>).toMatchFlightSnapshot();
-});
-
-it(`renders NativeTabsTriggerTabBar`, async () => {
-  await expect(<NativeTabsTriggerTabBar />).toMatchFlightSnapshot();
 });